requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.24.0
pytest>=7.0.0
hydra-core>=1.3.0
wandb>=0.15.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0 